# transportManager/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transport Management's API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/employer/otp": {
            "get": {
                "tags": [
                    "Employer"
                ],
                "summary": "otp code",
                "operationId": "ffb89d80835149f8eab5f185a6c754d5",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/employer/login": {
            "get": {
                "tags": [
                    "Employer"
                ],
                "summary": "login",
                "operationId": "35586c95fc4c8dd8b03ddfb19c6a647d",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/employer/register/person": {
            "post": {
                "tags": [
                    "Employer"
                ],
                "summary": "register",
                "operationId": "5a9855f344ed709108b3067386b96e82",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "national_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/employer/register/company": {
            "post": {
                "tags": [
                    "Employer"
                ],
                "summary": "register",
                "operationId": "ef81c7ea1c5ce6bf51ccd4b412e09705",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "economic_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "registration_number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "national_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/employer/ticket": {
            "get": {
                "tags": [
                    "Employer"
                ],
                "operationId": "f4a8290f2b92d3d11d4a8142df988ce4",
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employer"
                ],
                "operationId": "01ee8a918e01385b206d44abdac5a576",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "text"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Ticket title"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Ticket message"
                                    },
                                    "file": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/employer/ticket/{id}": {
            "get": {
                "tags": [
                    "Employer"
                ],
                "operationId": "329b26ba072e94d54a1f6bb8c178d369",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/employer/ticket/{ticket}": {
            "post": {
                "tags": [
                    "Employer"
                ],
                "operationId": "5c84dcf18af7780576f70c74d229f19f",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Ticket Reply"
                                    },
                                    "file": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}